requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
alabaster==1.0.0
annotated-types==0.7.0
applaunchservices==0.3.0
appnope==0.1.4
arrow==1.3.0
astroid==3.3.10
asttokens==3.0.0
asyncssh==2.21.0
atomicwrites==1.4.1
attrs==25.3.0
autopep8==2.0.4
babel==2.17.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
binaryornot==0.4.4
black==25.1.0
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
comm==0.2.2
cookiecutter==2.6.0
cryptography==45.0.4
dash==3.0.4
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
diff-match-patch==20241021
dill==0.4.0
docstring-to-markdown==0.17
docutils==0.21.2
et_xmlfile==2.0.0
executing==2.2.0
fastjsonschema==2.21.1
flake8==7.1.2
Flask==3.0.3
frozenlist==1.7.0
google-ai-generativelanguage==0.4.0
google-api-core==2.25.1
google-api-python-client==2.173.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.3.2
googleapis-common-protos==1.70.0
grpcio==1.73.0
grpcio-status==1.62.3
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
imagesize==1.4.1
importlib_metadata==8.7.0
inflection==0.5.1
intervaltree==3.1.0
ipykernel==6.29.5
ipython==8.37.0
isort==6.0.1
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.2
jellyfish==1.2.0
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
keyring==25.6.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mistune==3.1.3
more-itertools==10.7.0
multidict==6.5.0
mypy_extensions==1.1.0
narwhals==1.43.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==2.3.0
numpydoc==1.8.0
openpyxl==3.1.5
packaging==25.0
pandas==2.3.0
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pickleshare==0.7.5
platformdirs==4.3.8
plotly==6.1.2
pluggy==1.6.0
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==4.25.8
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydocstyle==6.3.0
pyflakes==3.2.0
PyGithub==2.6.1
Pygments==2.19.1
PyJWT==2.10.1
pylint==3.3.7
pylint-venv==3.0.4
pyls-spyder==0.4.0
PyNaCl==1.5.0
pyobjc-core==11.1
pyobjc-framework-Cocoa==11.1
pyobjc-framework-CoreServices==11.1
pyobjc-framework-FSEvents==11.1
pyparsing==3.2.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.17
PyQt5_sip==12.17.0
PyQtWebEngine==5.15.7
PyQtWebEngine-Qt5==5.15.17
python-dateutil==2.9.0.post0
python-lsp-black==2.0.0
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.12.2
python-slugify==8.0.4
pytoolconfig==1.3.1
pytz==2025.2
pyuca==1.2
PyYAML==6.0.2
pyzmq==27.0.0
QDarkStyle==3.2.3
qstylizer==0.2.4
QtAwesome==1.4.0
qtconsole==5.6.1
QtPy==2.4.3
referencing==0.36.2
requests==2.32.4
retrying==1.3.4
rich==14.0.0
roman-numerals-py==3.1.0
rope==1.13.0
rpds-py==0.25.1
rsa==4.9.1
rtree==1.4.0
setuptools==80.9.0
six==1.17.0
snowballstemmer==3.0.1
sortedcontainers==2.4.0
soupsieve==2.7
Sphinx==8.2.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
spyder==6.0.7
spyder-kernels==3.0.5
stack-data==0.6.3
superqt==0.7.5
tabulate==0.9.0
text-unidecode==1.3
textdistance==4.6.3
three-merge==0.1.1
tinycss2==1.4.0
tomlkit==0.13.3
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20250516
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ujson==5.10.0
uritemplate==4.2.0
urllib3==2.5.0
watchdog==6.0.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.6
whatthepatch==1.0.7
wrapt==1.17.2
wurlitzer==3.1.1
yapf==0.43.0
yarl==1.20.1
zipp==3.23.0
